requirements.txt
matplotlib~=3.7.1
numpy~=1.24.3
pandas~=1.5.3
scikit-learn~=1.2.2
pytorch~=1.13.1
torchmetrics~=0.11.4
